+++ src/locale/index.tsx
@@ -4,22 +4,15 @@
 import { AppVersion, DefaultLocale } from 'consts';
 import i18next from 'i18next';
 import { initReactI18next } from 'react-i18next';
-import { AnyJson } from 'types';
 import baseEn from './en/base.json';
 import helpEn from './en/help.json';
-import { doDynamicImport } from './utils';
+import { doDynamicImport, getActiveLanguage, getResources } from './utils';
 
-// the supported namespaces
+// the supported namespaces.
 export const lngNamespaces = ['base', 'help'];
 
-// the active language
-let lng: string;
-
-// whether a dynamic import is needed
-let dynamicLoad = false;
-
-// default structure of language resources
-const resourcesInnerDefault = { ...baseEn, ...helpEn };
+// default structure of language resources.
+const fallbackResources = { ...baseEn, ...helpEn };
 
 // check app version, wipe `lng_resources` if version is different.
 const localAppVersion = localStorage.getItem('app_version');
@@ -32,51 +25,12 @@ if (localAppVersion !== AppVersion) {
   localStorage.setItem('app_version', AppVersion);
 }
 
-// determine the default language.
-const localLng = localStorage.getItem('lng');
-if (!localLng) {
-  lng = DefaultLocale;
-  localStorage.setItem('lng', DefaultLocale);
-} else {
-  lng = localLng;
-}
+// get active language.
+const lng: string = getActiveLanguage();
 
-// determine resources exist without dynamically importing them.
-let resources: AnyJson = null;
-if (lng === DefaultLocale) {
-  resources = {
-    en: resourcesInnerDefault,
-  };
-  localStorage.setItem(
-    'lng_resources',
-    JSON.stringify({ l: DefaultLocale, r: resourcesInnerDefault })
-  );
-} else {
-  // not the default locale, check if local resources exist
-  let localValid = false;
-  const localResources = localStorage.getItem('lng_resources');
-  if (localResources !== null) {
-    const { l, r } = JSON.parse(localResources);
-
-    if (l === lng) {
-      localValid = true;
-      // local resources found, load them in
-      resources = {
-        [lng]: {
-          ...r,
-        },
-      };
-    }
-  }
-
-  if (!localValid) {
-    // no resources exist locally, dynamic import needed.
-    dynamicLoad = true;
-    resources = {
-      en: resourcesInnerDefault,
-    };
-  }
-}
+// get default resources and whether a dynamic load is required for
+// the active language.
+const { resources, dynamicLoad } = getResources(lng, fallbackResources);
 
 // configure i18n object.
 i18next.use(initReactI18next).init({
